重构 PdoStatement
