+++ src/Db/Mysql/Contract/IMysqlDb.php
@@ -10,6 +10,8 @@ interface IMysqlDb extends IDb
 {
     /**
      * 检查错误码是否为掉线
+     *
+     * @param int $code
      */
-    public function checkCodeIsOffline(int $code): bool;
+    public function checkCodeIsOffline($code): bool;
 }
